remove: remove old sheets state
