+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import { MacHeader } from "../components/headers/MacHeader";
 import { SearchInput } from "../components/inputs/SearchInput";
-import { useContinuousSync } from "../utils/hooks/useContinuousSync";
+// import { useContinuousSync } from "../utils/hooks/useContinuousSync";
 import { Loader } from "../components/svg/Loader";
 import { useIsClient } from "../utils/hooks/useIsClient";
 import {Sheet} from "../components/girds/Sheet";
