+++ src/state/sheets.ts
@@ -1,94 +1,114 @@
-import { create } from "zustand";
-import { persist } from "zustand/middleware";
-import { immer } from "zustand/middleware/immer";
-import { ICell, ISheet } from "../types/sheet";
-import { mockSheet } from "../assets/mock";
-import { computeCell } from "../utils/cells";
-import { getRowCells } from '../utils/sheet';
+import { notationToCoordinates } from '../utils/cells';
+import { types, Instance, IAnyModelType } from "mobx-state-tree";
+import { coordinatesToNotation } from '../utils/cells';
+import { cellContainsSearchValue } from '../utils/search';
 
-export interface ISheetsState {
-  selectedSheetId: string | null;
-  sheets: Record<string, ISheet>;
-  addNewRow: () => void;
-  removeRow: (row: number) => void;
-  updateSheet: (updates: Partial<ISheet>) => void;
-  updateCell: (cell: ICell) => void;
-  setCellEdit: (cell: ICell) => void;
+const CellModel = types
+  .model("Cell", {
+    row: types.number,
+    col: types.number,
+    edit: types.boolean,
+    value: types.string,
+    error: types.maybe(types.boolean),
+    message: types.maybe(types.string),
+    inputCells: types.maybe(types.array(types.late((): IAnyModelType => CellModel))),
+  })
+  .views(self => ({
+    get id() {
+      return coordinatesToNotation(self.row, self.col);
+    },
+    get computed() {
+      return self.value;
+    },
+  }))
+  .actions(self => ({
+    setValue(value: string) {
+      self.value = value;
+    },
+    setEdit(edit?: boolean) {
+      self.edit = !!edit;
+    },
+    containsSearchText(search?: string | null) {
+      if (typeof search == 'string') {
+        return cellContainsSearchValue(search, self.value, self.computed)
+      }
+      return false;
+    }
+  }));
+
+type CellType = Instance<typeof CellModel>;
+
+const ColHeaderModel = types
+  .model("ColHeader", {
+    col: types.number,
+    value: types.string,
+  });
+
+type ColHeaderType = Instance<typeof ColHeaderModel>;
+
+const SyncStatus = types.enumeration('SyncStatus', ['IN_PROGRESS', 'DONE', 'ERROR']);
+
+const SheetModel = types
+  .model("Sheet", {
+    id: types.identifier,
+    status: types.maybe(SyncStatus),
+    serverId: types.maybe(types.string),
+    rows: types.number,
+    cols: types.number,
+    title: types.maybe(types.string),
+    colNames: types.array(ColHeaderModel),
+    cells: types.array(types.array(CellModel)),
+    updatedAt: types.maybe(types.string),
+    savedAt: types.maybe(types.string),
+  })
+  .actions(self => ({
+    getCell(row: number, col: number) {
+      return self.cells[row]?.[col];
+    },
+    getCellById(id: string) {
+      const [row, col] = notationToCoordinates(id)
+      return self.cells[row]?.[col];
+    },
+    getRow(row: number) {
+      return self.cells[row];
+    }
+  }));
+
+type SheetType = Instance<typeof SheetModel>;
+
+const SheetsStateModel = types
+  .model("SheetsState", {
+    selectedSheetId: types.maybeNull(types.string),
+    sheets: types.map(SheetModel),
+  })
+  .views(self => ({
+    get selectedSheet() {
+      if (!self.selectedSheetId) {
+        return null;
+      }
+      return self.sheets.get(self.selectedSheetId)
+    }
+  }))
+  .actions(self => ({
+    setSelectedSheetId(id: string | null) {
+      self.selectedSheetId = id;
+    },
+    addSheet(sheet: SheetType) {
+      self.sheets.set(sheet.id, sheet);
+    },
+  }));
+
+type SheetsStateType = Instance<typeof SheetsStateModel>;
+
+export {
+  SheetModel,
+  ColHeaderModel,
+  SheetsStateModel
+}
+export type {
+  CellType,
+  SheetType,
+  ColHeaderType,
+  SheetsStateType,
 }
 
-export const useSheets = create(
-  persist(
-    immer<ISheetsState>(set => ({
-      selectedSheetId: mockSheet.id,
-      sheets: {
-        [mockSheet.id]: mockSheet,
-      },
-      updateSheet: (updates: Partial<ISheet>) => {
-        set((s) => {
-          const slectedId = s.selectedSheetId;
-          if (slectedId && s.sheets[slectedId]) {
-            s.sheets[slectedId] = {
-              ...s.sheets[slectedId],
-              ...updates
-            }
-          }
-        });
-      },
-      addNewRow: () => {
-        set((s) => {
-          const slectedId = s.selectedSheetId;
-          if (slectedId && s.sheets[slectedId]) {
-            s.sheets[slectedId].rows += 1;
-            s.sheets[slectedId].updatedAt = new Date().toISOString();
-          }
-        })
-      },
-      removeRow: (row: number) => {
-        set((s) => {
-          const slectedId = s.selectedSheetId;
-          if (slectedId && s.sheets[slectedId]) {
-            const cellsToRemove = getRowCells(row, s.sheets[slectedId].cols);
-            alert(`TODO: Delete ${cellsToRemove} :))))`)
-          }
-        });
-      },
-      setCellEdit: (cell: ICell) => {
-        set((s) => {
-          const slectedId = s.selectedSheetId;
-          if (slectedId && s.sheets[slectedId]) {
-            s.sheets[slectedId].data[cell.id] = cell;
-          }
-        });
-      },
-      updateCell: (cell: ICell) => {
-        set((s) => {
-          const slectedId = s.selectedSheetId;
-          if (slectedId && s.sheets[slectedId]) {
-            const { id } = cell;
-            const cells = s.sheets[slectedId].data;
-            const { computed, inputCellRefsToRemove } = computeCell(cell, cells);
-            s.sheets[slectedId].data[id] = computed;
-
-            // Add cell references to input cells
-            Object.keys(computed.inputCells).forEach(cellId => {
-              s.sheets[slectedId].data[cellId].outputCells[id] = true;
-            });
-            // Remove cell reference from no longed dependeant cells
-            inputCellRefsToRemove.forEach(cellId => {
-              delete s.sheets[slectedId].data[cellId].outputCells[id];
-            });
-            // Recalculate depending cells
-            Object.keys(computed.outputCells).forEach(cellId => {
-              const { computed } = computeCell(s.sheets[slectedId].data[cellId], cells);
-              s.sheets[slectedId].data[cellId] = computed;
-            });
-            s.sheets[slectedId].updatedAt = new Date().toISOString();
-          }
-        });
-      },
-    })),
-    {
-      name: "stacking-storage",
-    }
-  )
-);
